requirements.txt
absl-py==2.1.0
altair==5.3.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
coverage==7.4.3
cycler==0.12.1
data==0.4
DateTime==5.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
funcsigs==1.0.2
future==1.0.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.62.2
grpcio-status==1.62.2
h5py==3.10.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.3.2
kiwisolver==1.4.5
latex==0.7.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.3
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.2
panda==0.3.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.22.0
plotly-express==0.4.1
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pvlib==0.10.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
SciencePlots==2.1.1
scikeras==0.13.0
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
shutilwhich==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.35.0
streamlit-nightly==1.35.1.dev20240531
sympy==1.12
tempdir==0.7.1
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==2.1.2
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.9
wrapt==1.16.0
zope.interface==6.4.post2
